+++ web/pwa/src/app/model/Kvs.ts
@@ -19,4 +19,5 @@ export interface Kvs {
     enableRemoteInstall: boolean;
     enableAudioOnFirstInteraction: boolean;
     snapshotIntegrityCheck: boolean;
+    infoId?: number;
 }